T337002 | Move helper functions to a separate class
